Follow writing exporters guidelines more closely
This commit tries make this exporter adhere more closely to the official
"writing exporters" document. Most importantly it changes the direct
instrumentation approach to creating new const metrics. It also changes
to build toolchain to use the promu tool. This exporter still needs to
use caching of scrape results contrary to the recommended best practice
as some of the queries may not run as often as prometheus scrapes this
exporter.