feat: allow disable built in metrics
This PR adds the `-metrics.disable-built-in` flag in order to disable
the collection and export of built-in metrics, such as GC statistics.

In some cases this is desirable. In many cases I would suspect this
would be a good default as in reality I am not concerned with how well
the GC is working on a tiny app running SQL queries every few minutes.